Fix: OTP verification, updated file names
