+++ src/Pattern/Assets/OTP_VERIFICATION.ts
@@ -1,0 +1,41 @@
+export const CONFIRM_CLOSE = [
+  {
+    src: new URL('~/src/AssetFiles/illustration.svg', import.meta.url).href,
+    alt: 'Cancel Trancation',
+    as: 'image/svg'
+  }
+]
+
+export const SUCCESS = [
+  {
+    src: new URL('~/src/AssetFiles/successTick.svg', import.meta.url).href,
+    alt: 'success ',
+    as: 'image/svg'
+  }
+]
+
+export const BANK_LOGO = [
+  {
+    src: new URL('~/src/AssetFiles/bank.svg', import.meta.url).href,
+    alt: 'bank logo ',
+    as: 'image/svg'
+  }
+]
+
+export const OTP_EXHAUST = [
+  {
+    src: new URL('~/src/AssetFiles/otpExhaust.svg', import.meta.url).href,
+    alt: 'bank logo ',
+    as: 'image/svg'
+  }
+]
+
+export const SOMETHING_WENT_WRONG = [
+  //
+  {
+    src: new URL('~/src/AssetFiles/somethingWentWrong.svg', import.meta.url)
+      .href,
+    alt: 'something went wrong',
+    as: 'image/svg'
+  }
+]
